commands: Update override parsing tests
