+++ tests/config/config_test.py
@@ -106,31 +106,3 @@ class ValidateTest(unittest.TestCase):
         self.assertEqual({'foo': {'bar': 'bad'}}, errors)
 
     # TODO: add more tests
-
-
-class ParseOverrideTest(unittest.TestCase):
-    def test_valid_override(self):
-        expected = (b'section', b'key', b'value')
-        self.assertEqual(expected, config.parse_override(b'section/key=value'))
-        self.assertEqual(
-            expected, config.parse_override(b'section/key=value '))
-        self.assertEqual(
-            expected, config.parse_override(b'section/key =value'))
-        self.assertEqual(
-            expected, config.parse_override(b'section /key=value'))
-
-    def test_valid_override_is_bytes(self):
-        section, key, value = config.parse_override(b'section/key=value')
-        self.assertIsInstance(section, bytes)
-        self.assertIsInstance(key, bytes)
-        self.assertIsInstance(value, bytes)
-
-    def test_empty_override(self):
-        expected = ('section', 'key', '')
-        self.assertEqual(expected, config.parse_override(b'section/key='))
-        self.assertEqual(expected, config.parse_override(b'section/key=  '))
-
-    def test_invalid_override(self):
-        self.assertRaises(ValueError, config.parse_override, b'section/key')
-        self.assertRaises(ValueError, config.parse_override, b'section=')
-        self.assertRaises(ValueError, config.parse_override, b'section')
